Extract HttpCache + Interceptor to public package
