+++ apollo-runtime/src/main/java/com/apollographql/apollo/cache/http/HttpCacheInterceptor.java
@@ -1,4 +1,4 @@
-package com.apollographql.apollo.internal.cache.http;
+package com.apollographql.apollo.cache.http;
 
 import com.apollographql.apollo.internal.util.ApolloLogger;
 
@@ -9,20 +9,19 @@ import okhttp3.Request;
 import okhttp3.Response;
 
 import static com.apollographql.apollo.api.internal.Utils.checkNotNull;
-import static com.apollographql.apollo.internal.cache.http.HttpCache.CACHE_KEY_HEADER;
-import static com.apollographql.apollo.internal.cache.http.Utils.closeQuietly;
-import static com.apollographql.apollo.internal.cache.http.Utils.isNetworkFirst;
-import static com.apollographql.apollo.internal.cache.http.Utils.isNetworkOnly;
-import static com.apollographql.apollo.internal.cache.http.Utils.isPrefetchResponse;
-import static com.apollographql.apollo.internal.cache.http.Utils.isStale;
-import static com.apollographql.apollo.internal.cache.http.Utils.shouldExpireAfterRead;
-import static com.apollographql.apollo.internal.cache.http.Utils.shouldSkipCache;
-import static com.apollographql.apollo.internal.cache.http.Utils.shouldSkipNetwork;
-import static com.apollographql.apollo.internal.cache.http.Utils.strip;
-import static com.apollographql.apollo.internal.cache.http.Utils.unsatisfiableCacheRequest;
-import static com.apollographql.apollo.internal.cache.http.Utils.withServedDateHeader;
+import static com.apollographql.apollo.cache.http.HttpCache.CACHE_KEY_HEADER;
+import static com.apollographql.apollo.cache.http.Utils.isNetworkFirst;
+import static com.apollographql.apollo.cache.http.Utils.isNetworkOnly;
+import static com.apollographql.apollo.cache.http.Utils.isPrefetchResponse;
+import static com.apollographql.apollo.cache.http.Utils.isStale;
+import static com.apollographql.apollo.cache.http.Utils.shouldExpireAfterRead;
+import static com.apollographql.apollo.cache.http.Utils.shouldSkipCache;
+import static com.apollographql.apollo.cache.http.Utils.shouldSkipNetwork;
+import static com.apollographql.apollo.cache.http.Utils.strip;
+import static com.apollographql.apollo.cache.http.Utils.unsatisfiableCacheRequest;
+import static com.apollographql.apollo.cache.http.Utils.withServedDateHeader;
 
-final class HttpCacheInterceptor implements Interceptor {
+public final class HttpCacheInterceptor implements Interceptor {
   private final HttpCache cache;
   private final ApolloLogger logger;
 
@@ -168,7 +167,7 @@ final class HttpCacheInterceptor implements Interceptor {
     }
 
     if (isStale(request, cachedResponse)) {
-      closeQuietly(cachedResponse);
+      Utils.closeQuietly(cachedResponse);
       return null;
     }
